+++ src/utils/buildNodeMaterialFromExisting.js
